backend/requirements.txt
fastapi
uvicorn
firebase-admin
google-generativeai
httpx
pytest

